fastapi-aossample-codespaces/requirements.txt
fastapi==0.112.0
uvicorn==0.30.0
pydantic==2.8.2
pytest==8.3.1
httpx==0.27.0
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
python-multipart==0.0.6
